requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.3
daphne==4.0.0
Deprecated==1.2.13
Django==4.1.3
faunadb==4.3.1
future==0.18.2
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iso8601==1.1.0
msgpack==1.0.4
packaging==21.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.6
redis==4.3.4
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.12
wrapt==1.14.1
zope.interface==5.5.1
